A permitted send says so, quietly
SendPermission drew nothing when the engine allowed the message. That was
deliberate — the overwhelming majority of sends are allowed and none
should cost attention — and the file's own header already called that
state "one quiet line".

But silence was also what the composer showed while it was still asking,
and before anybody had asked at all. Three situations, one blank space,
and no way to tell "checked and fine" from "not asked yet".

SendMark is the smallest thing that separates them: a CommunicationStatus
in the footer's lead slot, beside the controls a rep is about to press.
The Callout stays in the body, where a refusal that needs reading belongs.
A rep who never looks at the mark loses nothing, which is the test a
quiet signal has to pass.

It draws NOTHING before a recipient exists. A mark over an empty form
would claim an answer about a message nobody has addressed.

No new prop on ConfirmModal: the footer already has an actionsLead slot,
and twenty-eight surfaces build on that dialog. My first attempt invented
a confirmPrefix prop for a component that did not need one.

The mark cost fourteen lines in a file frozen at its count, so
composedraftband.tsx comes out to pay for it — the AI disclosure band and
the rewrites under it, which belong together because every rewrite asks
the same machine to write the words the band discloses. 2941 down to
2797, waiver ratcheted.

Two tests, one mutation-verified.
