+++ frontend/src/screens/compose.tsx
@@ -1,23 +1,11 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { Sparkles } from "lucide-react";
-import {
-  type ReactNode,
-  useCallback,
-  useEffect,
-  useId,
-  useRef,
-  useState,
-} from "react";
+import { useCallback, useEffect, useId, useRef, useState } from "react";
 import { api } from "../api/client";
 import type { components } from "../api/schema";
-import { useRecordZone } from "../app/recordzone";
 import { navigate } from "../app/router";
-import { Badge, Button } from "../design-system/atoms";
+import { Button } from "../design-system/atoms";
 import { Callout } from "../design-system/callout";
 import { ConfirmModal } from "../design-system/confirmmodal";
-import { Eyebrow } from "../design-system/eyebrow";
-import { OpenEmailDrawer } from "../design-system/openemaildrawer";
-import { Panel, PanelBody } from "../design-system/panel";
 import {
   liveProjects,
   type PickableProject,
@@ -29,14 +17,9 @@ import { paragraphsFrom, RichText } from "../design-system/richtext";
 import { Select } from "../design-system/select";
 import { useToast } from "../design-system/toast";
 import type { TokenSuggestion } from "../design-system/tokeninput";
-import {
-  formatDateTime,
-  INTL_LOCALE,
-  identifierNumber,
-} from "../format/format";
+import { formatDateTime, INTL_LOCALE } from "../format/format";
 import { viewerZone } from "../format/timezone";
 import { type Locale, useLocale, usePlural, useT } from "../i18n";
-import type { MessageKey } from "../i18n/en";
 import { entityTimelineKeys } from "./activitykeys";
 import {
   isConsentNotGranted,
@@ -47,7 +30,7 @@ import {
   useViewerId,
 } from "./common";
 import { recordNamesIn, useOrganization360 } from "./company360";
-import { StaleThreadNotice, VoiceDegradedNotice } from "./compose.notices";
+import { StaleThreadNotice } from "./compose.notices";
 import {
   asksWhy,
   type CommunicationContext,
@@ -61,11 +44,10 @@ import {
   type ChosenFile,
   useCarriageBlocks,
 } from "./composeattachments";
+import { DraftBand, RewriteRow } from "./composedraftband";
 import {
   AccountDraftContext,
   DraftOffer,
-  DraftReasons,
-  openCited,
   type PendingAction,
 } from "./composedraftcontext";
 import {
@@ -93,7 +75,6 @@ import {
   useThreadMessages,
 } from "./composethread";
 import { useRoster } from "./entityref";
-import { useOpenEmail } from "./openemail";
 import { usePerson360 } from "./person360";
 import type { Transport } from "./persontransports";
 import {
@@ -104,7 +85,7 @@ import {
   withSubjectTag,
 } from "./projectrecord";
 import { SCHEDULED_SCREEN } from "./scheduledsends";
-import { SendPermission } from "./sendpermission";
+import { SendMark, SendPermission } from "./sendpermission";
 import { useSendPermission } from "./usesendpermission";
 import { useVoiceProfile } from "./voice-profile";
 import "./compose.css";
@@ -1052,158 +1033,6 @@ export function missingToSend(
     missing.push("context");
   }
   return missing;
-}
-
-// The card that says a MACHINE wrote the words below, and what it wrote them
-// from: to a reader the Art. 50 disclosure and the draft's reasoning are one
-// statement — not your colleague's message, and here is what it stands on.
-//
-// `Panel tone="ai"` draws it, in the colour every other machine-authored
-// surface wears, its title at h3 under the drawer's own h2. Loudest thing in
-// the drawer on purpose: miss it and a model's words go out in a rep's name.
-//
-// The server's disclosure line is a compliance string rendered verbatim, never
-// reworded; a response that omits it still discloses, because a missing line
-// may not silently become a missing disclosure.
-//
-// The voice tag names the PROFILE version that styled the draft; the
-// provisional label reports what that profile is today. Neither implies a
-// weaker draft — nothing gates drafting on maturity. Both hang off the SERVED
-// version, because reporting a maturity over a draft no voice touched would
-// overstate this surface's own provenance, which Art. 50 does not permit.
-function DraftBand({
-  provenance,
-  maturity,
-  reasons,
-  children,
-}: Readonly<{
-  provenance: DraftProvenance;
-  maturity: VoiceProfile["maturity"] | undefined;
-  reasons: components["schemas"]["AccountDraftReason"][];
-  // The steer and the verb that asks for another draft: the card is the
-  // machine's own block, so asking it to write again belongs inside it.
-  children: ReactNode;
-}>) {
-  const t = useT();
-  // One drawer per CARD rather than per reason: a drawer per row would be
-  // several dialogs racing to be the one on top.
-  const [openEmail, setOpenEmail] = useOpenEmail();
-  const zone = useRecordZone();
-  if (!provenance.ai_generated) {
-    return null;
-  }
-  return (
-    <Panel
-      tone="ai"
-      title={t("compose.aiDisclosureTitle")}
-      titleLevel={3}
-      className="compose-band"
-    >
-      <PanelBody className="compose-band-body">
-        <p className="t-body">
-          {provenance.ai_disclosure || t("compose.aiDisclosureFallback")}
-        </p>
-        <DraftReasons
-          reasons={reasons}
-          onOpenRecord={openCited}
-          onOpenEmail={setOpenEmail}
-        />
-        <OpenEmailDrawer
-          activityId={openEmail}
-          zone={zone}
-          onClose={() => setOpenEmail(null)}
-        />
-        <VoiceDegradedNotice degraded={provenance.voice_degraded} />
-        {provenance.voice_profile_version != null && (
-          <>
-            <p className="t-caption">
-              {/* A profile VERSION, never grouped: version 1234 is one
-                  identifier, and "1.234" reads as a different one. */}
-              {t("compose.voiceVersion", {
-                n: identifierNumber(provenance.voice_profile_version),
-              })}
-            </p>
-            {maturity === "provisional" && (
-              <p className="t-caption">
-                <Badge>{t("compose.provisional")}</Badge>{" "}
-                {t("compose.provisionalHint")}
-              </p>
-            )}
-          </>
-        )}
-        {children}
-      </PanelBody>
-    </Panel>
-  );
-}
-
-// The four things a rep asks the machine to do to its own draft, as one press
-// each. Each is an instruction for ONE call — it never becomes the standing
-// steer in the intent field.
-// The label a rep reads and the instruction the model is given are two
-// different strings and both are translated: the button says "Shorter" and the
-// model is asked for it in a sentence, because an instruction of one word is
-// one the model has to guess the scope of.
-const REWRITES = [
-  {
-    key: "shorter",
-    label: "compose.rewriteShorter",
-    instruction: "compose.rewriteShorterAsk",
-  },
-  {
-    key: "warmer",
-    label: "compose.rewriteWarmer",
-    instruction: "compose.rewriteWarmerAsk",
-  },
-  {
-    key: "formal",
-    label: "compose.rewriteFormal",
-    instruction: "compose.rewriteFormalAsk",
-  },
-  {
-    key: "deadline",
-    label: "compose.rewriteDeadline",
-    instruction: "compose.rewriteDeadlineAsk",
-  },
-] as const satisfies readonly {
-  key: string;
-  label: MessageKey;
-  instruction: MessageKey;
-}[];
-
-// Offered only over the machine's OWN untouched words. Once the rep has
-// edited the body, a rewrite would throw their work away to answer a question
-// about text that is no longer there — so the row withdraws rather than
-// growing a confirm nobody would read.
-// `disabled` and not `pending`: these buttons do not report a write of their
-// own, they refuse to start a second one. Named for what it does, because named
-// for a state it does not have it read as the draft button's own spinner and a
-// change to that button silently unblocked these.
-function RewriteRow({
-  onRewrite,
-  disabled,
-}: Readonly<{
-  onRewrite: (instruction: string) => void;
-  disabled: boolean;
-}>) {
-  const t = useT();
-  return (
-    <div className="compose-rewrite">
-      <Eyebrow>{t("compose.rewrite")}</Eyebrow>
-      {REWRITES.map((rewrite) => (
-        <Button
-          key={rewrite.key}
-          small
-          variant="aiQuiet"
-          disabled={disabled}
-          onClick={() => onRewrite(t(rewrite.instruction))}
-        >
-          <Sparkles aria-hidden="true" />
-          {t(rewrite.label)}
-        </Button>
-      ))}
-    </div>
-  );
 }
 
 // The mail-only half of the composer: AI drafting (there is no draft-message
@@ -2560,20 +2389,34 @@ export function ComposeModal({
         }}
         pending={send.isPending}
         error={sendError}
+        // The mark leads the footer row, before discard and the send controls:
+        // it is what a rep checks before pressing anything. The Callout in the
+        // body explains a refusal; this says the engine looked and found
+        // nothing wrong, which the composer used to say by drawing nothing —
+        // indistinguishable from not having asked at all.
         actionsLead={
-          discardControl && (
-            <Button
-              onClick={discardControl.run}
-              disabled={discardControl.disabled}
-              // What discarding DOES, on the control itself: it is not an undo,
-              // it tells the voice profile this draft missed. A rep who reads it
-              // as "clear the box" would train the model on every draft they
-              // merely changed their mind about.
-              title={t("compose.discardDraftHint")}
-            >
-              {t("compose.discardDraft")}
-            </Button>
-          )
+          <>
+            {!isChannelReply && (
+              <SendMark
+                preview={permission.preview}
+                asking={permission.asking}
+                unanswered={permission.unanswered}
+              />
+            )}
+            {discardControl && (
+              <Button
+                onClick={discardControl.run}
+                disabled={discardControl.disabled}
+                // What discarding DOES, on the control itself: it is not an undo,
+                // it tells the voice profile this draft missed. A rep who reads it
+                // as "clear the box" would train the model on every draft they
+                // merely changed their mind about.
+                title={t("compose.discardDraftHint")}
+              >
+                {t("compose.discardDraft")}
+              </Button>
+            )}
+          </>
         }
         confirmMenu={
           isChannelReply ? undefined : (
